feat(featgen): generate model-layer constants from YAML
Add FeatureType and EndpointType code generation to mash-featgen,
driven by feature_types and endpoint_types registries in
protocol-versions.yaml. This replaces hand-written constants with
generated _gen.go files while keeping hand-written helpers
(FeatureVendorBase, IsVendorFeature, FeatureMapBit, IsACEndpoint,
IsDCEndpoint, SupportsEnergyControl) in place.

- Add RawModelTypeDef struct and extend RawProtocolVersion
- Add GenerateFeatureTypes/GenerateEndpointTypes with templates
- Add ValidateFeatureIDs for registry/YAML consistency checks
- Add -model-output flag to mash-featgen CLI
- Add toEndpointGoName helper with HVAC initialism support
- Add feature YAML definitions and shared types
- Generate feature code from YAML via mash-featgen
- Strip hand-written type declarations from model package

+++ mash-go/pkg/cert/types.go
@@ -146,7 +146,7 @@ type DeviceIdentity struct {
 	VendorID uint32
 
 	// ProductID identifies the device product within the vendor.
-	ProductID uint32
+	ProductID uint16
 
 	// SerialNumber is the device serial number (optional).
 	SerialNumber string
